fix to report_alignment
This script is designed to work directly with strings, thus does not need pre-numerization. The issue was that report_alignment() funciton was derived from JIT version and was still  trying to de-numerize the alignments. It is fixed now.
Please note that this version does not benefit from cuda cores and JIT and thus slower.